pyproject.toml
[tool.poetry]
name = "mvp-liquid"
version = "0.1.0"
description = ""
authors = ["WrldEngine <kamscienceit123.kamran@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.1"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
black = "^24.4.2"
httpx = "^0.27.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.8.2"}
python-multipart = "^0.0.9"
celery = {extras = ["redis"], version = "^5.4.0"}
fastapi-mail = "^1.4.1"
tronpy = "^0.5.0"
bip32utils = "^0.3.post4"
mnemonic = "^0.21"
web3 = "^6.20.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
